+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/utils/FlinkRelOptUtil.scala
@@ -83,8 +83,68 @@ object FlinkRelOptUtil {
       detailLevel,
       withIdPrefix,
       withRetractTraits,
-      withRowType)
+      withRowType,
+      withTreeStyle = true)
     rel.explain(planWriter)
+    sw.toString
+  }
+
+  /**
+    * Gets the digest for a rel tree.
+    *
+    * The digest of RelNode should contain the result of RelNode#explain method, retraction traits
+    * (for StreamPhysicalRel) and RelNode's row type.
+    *
+    * Row type is part of the digest for the rare occasion that similar
+    * expressions have different types, e.g.
+    * "WITH
+    * t1 AS (SELECT CAST(a as BIGINT) AS a, SUM(b) AS b FROM x GROUP BY CAST(a as BIGINT)),
+    * t2 AS (SELECT CAST(a as DOUBLE) AS a, SUM(b) AS b FROM x GROUP BY CAST(a as DOUBLE))
+    * SELECT t1.*, t2.* FROM t1, t2 WHERE t1.b = t2.b"
+    *
+    * the physical plan is:
+    * {{{
+    *  HashJoin(where=[=(b, b0)], join=[a, b, a0, b0], joinType=[InnerJoin],
+    *    isBroadcast=[true], build=[right])
+    *  :- HashAggregate(groupBy=[a], select=[a, Final_SUM(sum$0) AS b])
+    *  :  +- Exchange(distribution=[hash[a]])
+    *  :     +- LocalHashAggregate(groupBy=[a], select=[a, Partial_SUM(b) AS sum$0])
+    *  :        +- Calc(select=[CAST(a) AS a, b])
+    *  :           +- ScanTable(table=[[builtin, default, x]], fields=[a, b, c])
+    *  +- Exchange(distribution=[broadcast])
+    *     +- HashAggregate(groupBy=[a], select=[a, Final_SUM(sum$0) AS b])
+    *        +- Exchange(distribution=[hash[a]])
+    *           +- LocalHashAggregate(groupBy=[a], select=[a, Partial_SUM(b) AS sum$0])
+    *              +- Calc(select=[CAST(a) AS a, b])
+    *                 +- ScanTable(table=[[builtin, default, x]], fields=[a, b, c])
+    * }}}
+    *
+    * The sub-plan of `HashAggregate(groupBy=[a], select=[a, Final_SUM(sum$0) AS b])`
+    * are different because `CAST(a) AS a` has different types, where one is BIGINT type
+    * and another is DOUBLE type.
+    *
+    * If use the result of `RelOptUtil.toString(aggregate, SqlExplainLevel.DIGEST_ATTRIBUTES)`
+    * on `HashAggregate(groupBy=[a], select=[a, Final_SUM(sum$0) AS b])` as digest,
+    * we will get incorrect result. So rewrite `explain_` method of `RelWriterImpl` to
+    * add row-type to digest value.
+    *
+    * @param rel rel node tree
+    * @return The digest of given rel tree.
+    */
+  def getDigest(rel: RelNode): String = {
+    val sw = new StringWriter
+    rel.explain(new RelTreeWriterImpl(
+      new PrintWriter(sw),
+      explainLevel = SqlExplainLevel.DIGEST_ATTRIBUTES,
+      // ignore id, only contains RelNode's attributes
+      withIdPrefix = false,
+      // add retraction traits to digest for StreamPhysicalRel node
+      withRetractTraits = true,
+      // add row type to digest to avoid corner case that similar
+      // expressions have different types
+      withRowType = true,
+      // ignore tree style, only contains RelNode's attributes
+      withTreeStyle = false))
     sw.toString
   }
