+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/utils/RelTreeWriterImpl.scala
@@ -38,7 +38,8 @@ class RelTreeWriterImpl(
     explainLevel: SqlExplainLevel = SqlExplainLevel.EXPPLAN_ATTRIBUTES,
     withIdPrefix: Boolean = false,
     withRetractTraits: Boolean = false,
-    withRowType: Boolean = false)
+    withRowType: Boolean = false,
+    withTreeStyle: Boolean = true)
   extends RelWriterImpl(pw, explainLevel, withIdPrefix) {
 
   var lastChildren: Seq[Boolean] = Nil
@@ -55,11 +56,13 @@ class RelTreeWriterImpl(
     }
 
     val s = new StringBuilder
-    if (depth > 0) {
-      lastChildren.init.foreach { isLast =>
-        s.append(if (isLast) "   " else ":  ")
+    if (withTreeStyle) {
+      if (depth > 0) {
+        lastChildren.init.foreach { isLast =>
+          s.append(if (isLast) "   " else ":  ")
+        }
+        s.append(if (lastChildren.last) "+- " else ":- ")
       }
-      s.append(if (lastChildren.last) "+- " else ":- ")
     }
 
     if (withIdPrefix) {
@@ -112,18 +115,30 @@ class RelTreeWriterImpl(
     }
     pw.println(s)
     if (inputs.length > 1) inputs.toSeq.init.foreach { rel =>
-      depth = depth + 1
-      lastChildren = lastChildren :+ false
+      if (withTreeStyle) {
+        depth = depth + 1
+        lastChildren = lastChildren :+ false
+      }
+
       rel.explain(this)
-      depth = depth - 1
-      lastChildren = lastChildren.init
+
+      if (withTreeStyle) {
+        depth = depth - 1
+        lastChildren = lastChildren.init
+      }
     }
     if (!inputs.isEmpty) {
-      depth = depth + 1
-      lastChildren = lastChildren :+ true
+      if (withTreeStyle) {
+        depth = depth + 1
+        lastChildren = lastChildren :+ true
+      }
+
       inputs.toSeq.last.explain(this)
-      depth = depth - 1
-      lastChildren = lastChildren.init
+
+      if (withTreeStyle) {
+        depth = depth - 1
+        lastChildren = lastChildren.init
+      }
     }
   }
 }
