+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/plan/stream/sql/DagOptimizationTest.scala
@@ -326,7 +326,7 @@ class DagOptimizationTest extends TableTestBase {
     val appendSink3 = util.createAppendTableSink(Array("a"), Array(INT))
     util.writeToSink(table3, appendSink3, "appendSink3")
 
-    util.verifyPlan()
+    util.verifyPlanWithTrait()
   }
 
   @Test
